+++ src/index.ts
@@ -44,7 +44,7 @@ const main = async (): Promise<void> => {
     const program = new commander.Command("create-cli-application")
       .version("0.6.0")
       .arguments("<application-name>")
-      .usage(`${chalk.yellowBright("<application-name>")} [options]`)
+      .usage(`${chalk.blueBright("<application-name>")} [options]`)
       .action((name) => {
         applicationName = name;
       })
@@ -60,11 +60,11 @@ const main = async (): Promise<void> => {
       )
       .on("--help", () => {
         console.log(
-          `\nOnly ${chalk.yellowBright("<application-name>")} is required.`
+          `\nOnly ${chalk.blueBright("<application-name>")} is required.`
         );
         console.log(`\nIf you run into a problem, please open up a new issue:`);
         console.log(
-          `${chalk.cyan(
+          `${chalk.blueBright(
             "https://github.com/alexlee-dev/create-cli-application/issues/new"
           )}\n`
         );