+++ src/program.ts
@@ -8,14 +8,14 @@ import commander from "commander";
 export const handleIncorrectApplicationName = (program: commander.Command) => {
   console.error("\nPlease specify the name of your application:");
   console.log(
-    `\t${chalk.cyan(program.name())} ${chalk.green("<application-name>")}\n`
+    `\t${chalk.blueBright(program.name())} ${chalk.green("<application-name>")}\n`
   );
   console.log("For example:");
   console.log(
-    `  ${chalk.cyan(program.name())} ${chalk.green("my-cool-app")}\n`
+    `  ${chalk.blueBright(program.name())} ${chalk.green("my-cool-app")}\n`
   );
   console.log(
-    `Run ${chalk.cyan(`${program.name()} --help`)} to see all options.`
+    `Run ${chalk.blueBright(`${program.name()} --help`)} to see all options.`
   );
   // * Exit the program
   process.exit(1);