+++ src/template/ts/src/setup.ts
@@ -12,7 +12,7 @@ const setup = async (state: AppState): Promise<void> => {
     clear();
 
     console.log(
-      `Welcome to ${chalk.yellowBright(
+      `Welcome to ${chalk.blueBright(
         "___APP NAME___"
       )}! Let's walk you through the initial set up.\n`
     );
@@ -21,7 +21,7 @@ const setup = async (state: AppState): Promise<void> => {
       {
         type: "input",
         name: "specialKey",
-        message: `Please enter your ${chalk.yellowBright("special key")}:`,
+        message: `Please enter your ${chalk.blueBright("special key")}:`,
       },
     ]);
     // * Template value. Change to be required setup or don't set up at all :)