+++ src/template/js/src/menu.js
@@ -1,9 +1,9 @@
 import boxen from "boxen";
 import chalk from "chalk";
 import inquirer from "inquirer";
-import { titleScreen } from "pickitt";
 
 import { blankBoxenStyle } from "./constants";
+import { titleScreen } from './util'
 
 /**
  * Displays Main Menu to user.