+++ src/template/ts/src/menu.ts
@@ -1,9 +1,9 @@
 import boxen from "boxen";
 import chalk from "chalk";
 import inquirer from "inquirer";
-import { titleScreen } from "pickitt";
 
 import { blankBoxenStyle } from "./constants";
+import { titleScreen } from "./util";
 
 import { AppState, MenuAction } from "./types";
 
@@ -75,7 +75,10 @@ export const interpretMenuAction = async (state: AppState): Promise<void> => {
       about: async (state: AppState): Promise<void> => {
         await titleScreen("___APP NAME___");
         console.log(
-          boxen(chalk.blueBright(`Author: `) + "___AUTHOR NAME___", blankBoxenStyle)
+          boxen(
+            chalk.blueBright(`Author: `) + "___AUTHOR NAME___",
+            blankBoxenStyle
+          )
         );
 
         console.log("Press any key to return to Main Menu ...");