+++ src/util.ts
@@ -2,6 +2,7 @@ import chalk from "chalk";
 import { spawn } from "child_process";
 import fs from "fs-extra";
 import path from "path";
+import semver from "semver";
 import validateProjectName from "validate-npm-package-name";
 
 /**
@@ -19,12 +20,13 @@ export const executeCommand = async (
     const cp = spawn(command, args, options);
     cp.on("error", (err: Error) => {
       if (err) {
+        console.log({ command, args, options });
         reject(err.message);
       }
     });
     cp.on("exit", (code: number | null, signal) => {
       if (code !== 0) {
-        reject({ code, signal });
+        reject({ args, command, code, signal, options });
       }
       resolve();
     });
@@ -48,7 +50,11 @@ export const cleanupError = async (
     // * Application Directory
     const root = path.resolve(applicationName);
 
-    await executeCommand("rimraf", [root]);
+    const rootExists = await fs.pathExists(root);
+
+    if (rootExists) {
+      await executeCommand("rimraf", [root]);
+    }
   } catch (error) {
     throw new Error(error);
   }
@@ -102,6 +108,17 @@ export const validateApplicationName = (applicationName: any) => {
 
     console.log("");
     console.error("Please choose a different application name.");
+    process.exit(1);
+  }
+};
+
+export const verifyNodeVersion = (): void => {
+  if (!semver.satisfies(process.version, ">=10.0.0")) {
+    console.error(
+      chalk.red(`create-cli-application requires Node v10 or higher.`)
+    );
+    console.error(chalk.red(`You are running Node ${process.version}.`));
+    console.error(chalk.red(`Please update your version of Node.`));
     process.exit(1);
   }
 };