+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - CONTRIBUTING.md
+- GIF examples
+- Use of `update-notifier` to alert the user to a newer version
 
 ### Changed