+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.0] - 2020-05-29
 
+### ❗Update Notifier
+
 ### Added
 
 - CONTRIBUTING.md