+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Format of website in `author` field of `package.json`
+- Changed terminal colors to be more more legible on bright terminals
 
 ### Removed