+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add missing `rimraf` dependency
 - Fix installing devDependencies without `--save-dev`
+- Fix `slant` not being recognized by Figlet as `Slant` in fonts
 
 ## [0.5.0] - 2020-05-29