+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Warning about unusable Node version (less than v10.0.0)
+- Wrapper to check if directory exists before removing with `rimraf`
 
 ### Changed